add init module with proto_input_to_native function
