+++ rust/tw_bitcoin/src/entry.rs
@@ -1,4 +1,5 @@
 use crate::modules::signer::Signer;
+use crate::utils::proto_input_to_native;
 use crate::{Error, Result};
 use bitcoin::address::NetworkChecked;
 use std::borrow::Cow;
@@ -144,116 +145,10 @@ impl BitcoinEntry {
         let mut builder = TransactionBuilder::new();
         let mut dummy_claims = vec![]; // TODO
 
-        for input in proto.inputs {
-            match input.to_recipient {
-                Proto::mod_Input::OneOfto_recipient::builder(b) => {
-                    match b.variant {
-                        OneOfvariant::p2sh(_) => todo!(),
-                        OneOfvariant::p2pkh(pubkey) => {
-                            let pubkey = pubkey_from_raw(&pubkey).unwrap();
-
-                            let (utxo, arg) = UtxoBuilder::new()
-                                .prev_txid(input.txid.as_ref().try_into().unwrap())
-                                .prev_index(input.vout)
-                                .amount(input.value as i64) // TODO: Just use u64 to begin with?
-                                .p2pkh(pubkey.clone())
-                                .unwrap();
-
-                            let sig = Signature::from_bytes(&vec![1; 65]).unwrap();
-                            let claim = SpendingScriptBuilder::new().p2pkh(sig, pubkey).unwrap();
-
-                            builder = builder.push_input(utxo, arg);
-                            dummy_claims.push(claim);
-                        },
-                        OneOfvariant::p2wsh(_) => todo!(),
-                        OneOfvariant::p2wpkh(pubkey) => {
-                            let pubkey = pubkey_from_raw(&pubkey).unwrap();
-
-                            let (utxo, arg) = UtxoBuilder::new()
-                                .prev_txid(input.txid.as_ref().try_into().unwrap())
-                                .prev_index(input.vout)
-                                .amount(input.value as i64) // TODO: Just use u64 to begin with?
-                                .p2wpkh(pubkey.clone())
-                                .unwrap();
-
-                            let sig = Signature::from_bytes(&vec![1; 65]).unwrap();
-                            let claim = SpendingScriptBuilder::new().p2wpkh(sig, pubkey).unwrap();
-
-                            builder = builder.push_input(utxo, arg);
-                            dummy_claims.push(claim);
-                        },
-                        OneOfvariant::p2tr_key_path(payload) => {
-                            // TODO: Rename field to `pubkey`?
-                            let pubkey = pubkey_from_raw(&payload.public_key).unwrap();
-
-                            let (utxo, arg) = UtxoBuilder::new()
-                                .prev_txid(input.txid.as_ref().try_into().unwrap())
-                                .prev_index(input.vout)
-                                .amount(input.value as i64) // TODO: Just use u64 to begin with?
-                                .p2tr_key_path(pubkey)
-                                .unwrap();
-
-                            let sig = schnorr::Signature::from_bytes(&vec![1; 64]).unwrap();
-                            let claim = SpendingScriptBuilder::new().p2tr_key_path(sig).unwrap();
-
-                            builder = builder.push_input(utxo, arg);
-                            dummy_claims.push(claim);
-                        },
-                        OneOfvariant::p2tr_script_path(_) => todo!(),
-                        OneOfvariant::brc20_inscribe(payload) => {
-                            // TODO: Rename field to `pubkey`?
-                            let pubkey = pubkey_from_raw(&payload.inscribe_to).unwrap();
-                            let ticker = payload.ticker.to_string();
-                            let value = payload.transfer_amount.to_string();
-
-                            let (utxo, arg) = UtxoBuilder::new()
-                                .prev_txid(input.txid.as_ref().try_into().unwrap())
-                                .prev_index(input.vout)
-                                .amount(input.value as i64) // TODO: Just use u64 to begin with?
-                                .brc20_transfer(pubkey.clone(), ticker.clone(), value.clone())
-                                .unwrap();
-
-                            let sig = schnorr::Signature::from_bytes(&vec![1; 64]).unwrap();
-                            let claim = SpendingScriptBuilder::new()
-                                .brc20_transfer(sig, pubkey, ticker, value)
-                                .unwrap();
-
-                            builder = builder.push_input(utxo, arg);
-                            dummy_claims.push(claim);
-                        },
-                        OneOfvariant::ordinal_inscribe(_) => todo!(),
-                        OneOfvariant::None => todo!(),
-                    }
-                },
-                Proto::mod_Input::OneOfto_recipient::custom_script(payload) => {
-                    let script_pubkey = Script::from(payload.script_pubkey.to_vec());
-                    let script_sig = Script::from(payload.script_sig.to_vec());
-                    let mut witness = Witness::new();
-                    for item in payload.witness_items {
-                        witness.push_item(Script::from(item.to_vec()));
-                    }
-
-                    let (utxo, arg) = UtxoBuilder::new()
-                        .prev_txid(input.txid.as_ref().try_into().unwrap())
-                        .prev_index(input.vout)
-                        .amount(input.value as i64) // TODO: Just use u64 to begin with?
-                        // TODO: Signing method:
-                        .custom_script_pubkey(
-                            script_pubkey,
-                            tw_utxo::signing_mode::SigningMethod::Legacy,
-                        )
-                        .unwrap();
-
-                    let claim = SpendingScriptBuilder::custom_script_sig_witness(
-                        Some(script_sig),
-                        Some(witness),
-                    );
-
-                    builder = builder.push_input(utxo, arg);
-                    dummy_claims.push(claim);
-                },
-                Proto::mod_Input::OneOfto_recipient::None => todo!(),
-            };
+        for input in &proto.inputs {
+            let (utxo, arg, claim) = proto_input_to_native(input)?;
+            builder = builder.push_input(utxo, arg);
+            dummy_claims.push(claim);
         }
 
         let mut utxo_outputs = vec![];
