+++ rust/tw_bitcoin/src/lib.rs
@@ -4,6 +4,7 @@
 
 pub mod entry;
 pub mod modules;
+mod utils;
 
 use std::fmt::Display;
